[fix] fix exceptions, refactoring SingleResponseDto
