+++ Server/src/main/java/com/e1i5/mentaltal/exception/ErrorResponse.java
@@ -8,7 +8,7 @@ import javax.validation.ConstraintViolation;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
-
+@Getter
 public class ErrorResponse {
     private int status;
     private String message;
